FileCryptor_0.5
- functionalitate completa
- citire de la tastatura cu validare
- continut text informativ 
- functii de criptare/decriptare cu exceptii
- metoda de mentinere a consistentei datelor la criptare imbunatatita; am renuntat la retinerea pozitiei si am atribuit fie valoarea 0 sau 1 pentru criptarile prea mari cu ultimul bit 0, respectiv 1, fie valoarea 2 pentru criptarile care nu depasesc limita si astfel vor fi ignorate
- fisierul criptat va contine forma hash a fisierului original, valorile ultimului bit(Cand e cazul) a fiecarei criptari, si criptarea in sine
- decriptare realizata conform cerintei, utilizand metoda de mentinere a consistentei criptarii cand e necesar
- daca decriptarea are succes, afisam in consola si locatia fisierului decriptat
- imbunatatiri generale

+++ FileCryptor/FileCryptor.py
@@ -2,7 +2,6 @@
 import sys
 import os
 import hashlib
-import traceback
 
 
 def hash(filePath):
@@ -21,23 +20,22 @@ def hash(filePath):
 
 
 def crypt(file,pw):
-    try:
-        # if not os.path.exists(file):
-        #     print("Fisierul dorit nu exista.")
-        # if os.path.isdir(file):
-        #     print("Path-ul introdus corespunde unui director. Introduceti path-ul unui fisier.")
-        # if not os.path.isfile:
-        #     pass
-        # if os.path.getsize(file) == 0:
-        #     print("File is empty!") 
+    try: 
+        if not os.path.isfile(file):
+            raise Exception("Path-ul introdus nu corespunde unui fisier. Introduceti path-ul unui fisier.")
+        if os.path.getsize(file) == 0:
+            raise Exception("Fisierul introdus este gol. Introduceti date valide.")
+        
         data = open(file, 'rt', encoding="utf-8")
         k = 0
         t = 0
+        
         newfile = os.path.splitext(file)[0] + '_1' + os.path.splitext(file)[1]
         c = 1
         while os.path.isfile(newfile):
             c += 1
             newfile = os.path.splitext(file)[0] + '_' + str(c) + os.path.splitext(file)[1]
+            
         try:
             crypted = open(newfile, 'w+', encoding="utf-8")
             m = hash(file)
@@ -49,16 +47,14 @@ def crypt(file,pw):
                 if not char:
                     break
                 if(ord(char)+ord(pw[t%len(pw)])) > 255:
-                    lastbits += str(t)
                     lastbits += bin(ord(char)+ord(pw[t%len(pw)]))[-1]
-                    lastbits += ','
+                else:
+                    lastbits += '2'
                 t += 1
-            
-            lastbits = lastbits[:-1]
             crypted.write(lastbits)
             crypted.write('\n')
-            data.seek(0)
             
+            data.seek(0)
             while(1):
                 char = data.read(1)
                 if not char:
@@ -71,7 +67,6 @@ def crypt(file,pw):
                     newchar = chr(ord(char) + ord(pw[k%len(pw)]))
                 crypted.write(newchar)
                 k += 1
-            
             crypted.close()
             
             cryptedfile = file + '.crypted'
@@ -81,24 +76,21 @@ def crypt(file,pw):
                 cryptedfile = os.path.splitext(file)[0] + '_' + str(d) + os.path.splitext(file)[1] + '.crypted'
             os.rename(newfile, cryptedfile)
         except:
-            print('except general 2')
-            print(traceback.format_exc())
-    except FileNotFoundError:
-        print('Fisierul nu a fost gasit. Introduceti date valide.')
-        print(traceback.format_exc())
+            print('Au aparut probleme in timpul procesului de criptare. Introduceti date valide.')
+    except Exception as e:
+        print(e)
     except:
-        print('except general 1')
-        print(traceback.format_exc())
+        print('Au aparut probleme in timpul procesului de criptare. Introduceti date valide.')
 
 
 def decrypt(file,pw):
     try:
-        # if not os.path.isfile(file):
-        #     print('fisierul nu exista.')
-        # if os.path.splitext(file)[1] != '.crypted':
-        #     print('fisierul are formatul gresit (nu e .crypted la final)')
-        # elif os.path.splitext(os.path.splitext(file)[0])[1] != '.txt':
-        #     print('fisierul are formatul gresit (nu e .txt.crypted)')
+        if not os.path.isfile(file):
+            raise Exception("Path-ul introdus nu corespunde unui fisier. Introduceti path-ul unui fisier.")
+        if os.path.splitext(file)[1] != '.crypted':
+            raise Exception('Fisierul nu are formatul corespunzator unui fisier criptat(".txt.crypted"). Introduceti date valide.')
+        elif os.path.splitext(os.path.splitext(file)[0])[1] != '.txt':
+            raise Exception('Fisierul nu are formatul corespunzator unui fisier criptat(".txt.crypted"). Introduceti date valide.')
         newfile = os.path.splitext(os.path.splitext(file)[0])[0] + '_1' + os.path.splitext(os.path.splitext(file)[0])[1]
         goodfile = os.path.splitext(file)[0]
         
@@ -116,18 +108,26 @@ def decrypt(file,pw):
         try:
             crypted = open(goodfile, 'w+', encoding="utf-8")
             
+            for count, line in enumerate(data):
+                pass
+            if(count+1 < 3):
+                print(count)
+                print('1')
+                raise Exception("Fisierul criptat contine date invalide. Introduceti date valide.")
+            
+            data.seek(0)
             oghash = data.readline().strip()
             lastbits = data.readline().strip()
-            
-            newlastbits = lastbits.split(',')
-            newlastbits = {int(lb[:-1]): int(lb[-1]) for lb in lastbits.split(',')}
+            x = ''
+            if any(x not in set("012") for x in lastbits):
+                raise Exception("Fisierul criptat contine date invalide. Introduceti date valide.")
             
             while(1):
                 char = data.read(1)
                 if not char:
                     break
-                if k in newlastbits:
-                    newnumber = ord(char) * 2 + newlastbits[k] - ord(pw[k%len(pw)])
+                if(int(lastbits[k]) != 2):
+                    newnumber = ord(char) * 2 + int(lastbits[k]) - ord(pw[k%len(pw)])
                     newchar = chr(newnumber)
                 else:
                     newchar = chr(ord(char) - ord(pw[k%len(pw)]))
@@ -136,22 +136,33 @@ def decrypt(file,pw):
             data.close()
             os.remove(newfile)
             crypted.close()
+            
+            newhash = hash(goodfile)
+            if(newhash == oghash):
+                print('Decriptare realizata cu succes.')
+                print('Locatie fisier decriptat:', goodfile)
+            else:
+                print('Decriptare esuata. Introduceti date valide.')
+        except Exception as e:
+            print(e)
         except:
-            pass
+            print('A aparut o problema in timpul decriptarii. Introduceti date valide.')
+    except Exception as e:
+        print(e)
     except:
-        pass
+        print('A aparut o problema in timpul decriptarii. Introduceti date valide.')
 
 
 def execinput():
     # INPUT: FileCryptor.py crypt abc.exe my_pass ⇒ va crea fisierul abc.exe.crypted
-    # INPUT: FileCryptor.py decrypt abc.exe.cryped my_pass ⇒ va crea fișierul abc.exe
+    # INPUT: FileCryptor.py decrypt abc.exe.crypted my_pass ⇒ va crea fișierul abc.exe
     if (len(sys.argv) - 1) < 3:
         print('Sintaxa incorecta1. Utilizati formatul:\n')
         print('FileCryptor.py [crypt/decrypt] path_fisier parola_criptare')
     else:
-        if(sys.argv[1] == 'crypt'):
+        if(sys.argv[1].lower() == 'crypt'):
             crypt(sys.argv[2], sys.argv[3])
-        elif(sys.argv[1].lower == 'decrypt'):
+        elif(sys.argv[1].lower() == 'decrypt'):
             decrypt(sys.argv[2], sys.argv[3])
         else: 
             print('Sintaxa incorecta. Utilizati formatul:\n')
@@ -159,5 +170,4 @@ def execinput():
 
 
 if __name__ == "__main__":
-    crypt(os.path.join(os.getcwd(), "plaintext.txt"), "ÈÉÊË")
-    decrypt(os.path.join(os.getcwd(), "plaintext.txt.crypted"),"ÈÉÊË")
+    execinput()
